refactor: PostBlock 타입스크립트로 전환
